+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/environment/Vocabulary.java
@@ -76,6 +76,7 @@ public class Vocabulary {
 
     public static final String P_HAS_SIMPLE_LIST = ATTRIBUTE_BASE + "C-hasSimpleSequence";
     public static final String P_HAS_REFERENCED_LIST = ATTRIBUTE_BASE + "C-hasReferencedSequence";
+    public static final String P_HAS_RDF_COLLECTION = ATTRIBUTE_BASE + "C-hasRDFCollection";
     public static final String P_HAS_RDF_SEQ = ATTRIBUTE_BASE + "C-hasRDFSeq";
 
     public static final String P_T_LOCAL_DATE_ATTRIBUTE = ATTRIBUTE_BASE + "tLocalDate";